% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tee.R
\name{say}
\alias{say}
\title{Interject a Message within a magrittr/dplyr Chain}
\usage{
say(.data, .message)
}
\arguments{
\item{.data}{the lhs of the chain evaluation}

\item{.message}{a message to be output to the console}
}
\value{
outputs the incomming data transparently
}
\description{
Interject a Message within a magrittr/dplyr Chain
}

