% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tee.R
\name{tee}
\alias{tee}
\title{tee Function within a magrittr/dplyr Chain}
\usage{
tee(.data, .message, .expr, ...)
}
\arguments{
\item{.data}{the lhs of the chain evaluation}

\item{.message}{a message to be output to the console}

\item{.expr}{expression to be evaluated}

\item{...}{named arguments to pass to the evalutating function}
}
\value{
outputs the incomming data transparently
}
\description{
tee Function within a magrittr/dplyr Chain
}

